safari-600.7-branch/LayoutTests/fast/spatial-navigation/snav-unit-overflow-and-scroll-in-direction-expected.txt

Start
End
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS true is true
PASS gFocusedDocument.activeElement.getAttribute("id") is "end"
